conversion cast to fix build
The compile error was also mentioned here: [1]

```
| ../sources/bmcweb-1.0+git/src/ssl_key_handler.cpp:293:61: error: conversion from 'std::chrono::duration<long long int>::rep' {aka 'long long int'} to 'long int' may change value [-Werror=conversion]
|   293 |     X509_gmtime_adj(X509_getm_notAfter(x509), duration.count());
|       |                                               ~~~~~~~~~~~~~~^~
| cc1plus: all warnings being treated as errors
```

The relevant function signature only accepts 'long' anways, so assuming
that's what the authors intention was.

```
ASN1_TIME *X509_gmtime_adj(ASN1_TIME *s, long adj);
```

Looking at the value of `duration.count()` it is `315569520`
which is representable by a 32-bit signed integer type.

References:
[1] https://gerrit.openbmc.org/c/openbmc/openbmc/+/89780/comments/d1fbedfb_3eca650e

+++ src/ssl_key_handler.cpp
@@ -290,7 +290,8 @@ static std::string constructX509(const std::string& cn, OpenSSLEVPKey& pPrivKey)
     X509_gmtime_adj(X509_getm_notBefore(x509), 0);
     // Cert is valid for 10 years
     std::chrono::seconds duration = std::chrono::years(10);
-    X509_gmtime_adj(X509_getm_notAfter(x509), duration.count());
+    X509_gmtime_adj(X509_getm_notAfter(x509),
+                    static_cast<long>(duration.count()));
 
     // set the public key to the key we just generated
     if (!x509Obj.setPubkey(pPrivKey))
